Update with SW-compatible speaker wiring/PADs
The original version of the RD reproduced an error in the spec
that resulted in PAD0-3 being labelled incorrectly. This meant
that the example 'Speaker' reference part of the circuit didn't
work with the default micro:bit software pin allocations.

I was hoping to be able to update all the versions at once, but as
this bug is out there it's better to fix it now and do the others
ASAP.

This is a fix for the Altium parts of the RD